Refactor: 콜렉션 심플 리스트 공통 컴포넌트로 리팩토링
